chore: separate storage tests
