+++ .github/workflows/sub-validate.yml
@@ -33,22 +33,6 @@ jobs:
     environment:
       name: ${{ inputs.stage }}
       url: ${{ inputs.stage-url }}
-    services:
-      mongodb:
-        image: mongo
-        ports:
-          - 27017:27017
-      postgres:
-        image: postgres:16
-        env:
-          POSTGRES_HOST_AUTH_METHOD: trust
-        ports:
-          - 5432:5432
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
     steps:
       - uses: actions/checkout@v3
